Updated 100-atoi.c

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -32,12 +32,12 @@ int _atoi(char *s)
 		}
 	}
 	if (ns1[1] != '\0' && ns1[0] == '-')
-		return ((unsigned int)ns1);
+		return ((long long int)ns1);
 	if (ns1[1] != '\0' && ns1[0] == '+')
 	{
 		for (a = 1; ns1[a] != '\0'; a++)
 			ns2[a - 1] = ns1[a];
-		return ((unsigned int)ns2);
+		return ((long long int)ns2);
 	}
 	if (ns1[1] == '\0')
 		return (0);
